Fix attribute serialization
