+++ tests/const.py
@@ -55,6 +55,9 @@ TEST_ENTITY_STATE_ATTRIBUTES_INCLUDE = [
         "float": 123.456,
         "list": [1, 2, 3, 4],
         "set": {5, 5},
+        "tuple": (6, 6),
+        "list_of_tuples": [(7, 7), (8, 8)],
+        "list_of_sets": [{9, 9}, {10, 10}],
         "none": None,
         "Collision Test": "first value",
         "collision_test": "second value",
